Add time constants and standardize refresh interval (delta times)

+++ ut-robomaster/src/subsystems/turret/turret_motor.hpp
@@ -22,7 +22,7 @@ public:
     void initialize();
     void reset();
     void updateMotorAngle();
-    void setAngle(float desiredAngle, uint32_t dt);
+    void setAngle(float desiredAngle, float dt);
     float getAngle();
     float getSetpoint();
